Feature: Handle ROT13 op on Empty Strings

+++ 0x06-pointers_arrays_strings/100-rot13.c
@@ -10,7 +10,7 @@
  */
 char *rot13(char *s)
 {
-	if ( strlen(s) <=0 )
+	if ( (int)strlen(s) <=0 )
 	{
 		return;
 	}
